Armory Link
Added the ability to query data using an armory link in addition to the
regular method. The formatting is `!armory pve <armory-link>`.

This allows people who use an in-game add-on to get an armory link to
quickly query and show friends in their channel the displayed data. It
also helps with accented characters.

User suggestion via Reddit, discussed in #14