Active Spec
Added the ability to get the players active spec, meaning the top
portion of the bots response will now say “110 Horde Holy Priest” for
example. The function has been setup like others so it returns an
object with some data in it, something which can be extended rather
easily should the ned come about in the future.

Addresses issue #15