Show Spec, Formatting, Armory Linking & More (#17)
* Armory Link

Added the ability to query data using an armory link in addition to the
regular method. The formatting is `!armory pve <armory-link>`.

This allows people who use an in-game add-on to get an armory link to
quickly query and show friends in their channel the displayed data. It
also helps with accented characters.

User suggestion via Reddit, discussed in #14

* split_string to split_query

Changing the language on the function call for consistency.

* Active Spec

Added the ability to get the players active spec, meaning the top
portion of the bots response will now say “110 Horde Holy Priest” for
example. The function has been setup like others so it returns an
object with some data in it, something which can be extended rather
easily should the ned come about in the future.

Addresses issue #15

* Test all the things

Added unit tests for talent specs, url and non-url queries.

* Spacing

Added some spacing.

* Cutting Edge Tracking

Added the ability to track Cutting Edge achievements. Now the AOTC or
CE achievement will only show if the character has completed, and only
the highest tier (for example CE will show instead of AOTC) completed.

This address #16

* Checking Against Criteria

For PVE achievements I have updated the bot so it now tracks against
criteria. Because PVP achievements don’t have unique criteria id’s for
some of the RBG achievements I’ve decided to keep these stats primarily
account-wide, whereas PVE is the opposite.

I feel that being able to tell the difference between an alt for PVE is
much more valuable than PVP.

#16

* Revert "Checking Against Criteria"

This reverts commit 0be8823c6e2d93b2e07818869eae65e082f684d5.

* Updating Query